Move logic from deduplicator into usecases
